Add Post Article Using WP Rest API
